+++ src/bpInfo.ts
@@ -1,1 +1,1 @@
-export const bpInfo = {version: 'v3.6.23'};
+export const bpInfo = {version: 'v3.6.24'};